+++ apps/server/src/mode-config.ts
@@ -1,0 +1,7 @@
+import dotenv from 'dotenv';
+
+export const MODE = process.env.MODE;
+
+dotenv.config({
+  path: MODE ? `.env.${MODE}` : '.env',
+});