+++ apps/web/src/util/api-interface.ts
@@ -1,7 +1,8 @@
 import { ApiInterface } from '@cbdc-markka/utils-api';
 
+const MODE = import.meta.env.MODE;
 export const apiInterface = new ApiInterface(
-  import.meta.env.PROD
+  MODE === 'production'
     ? 'https://cbdc-markka-server.onrender.com/api'
-    : 'http://localhost:3000/api',
+    : 'http://localhost:3001/api',
 );