+++ CHANGELOG.md
@@ -1,10 +1,11 @@
-## [2.1.0] - 10-03-2026
+## [2.1.0] - 11-03-2026
 
 ### Added
 
 - Caching for the user's session and account, used when the server doesn't respond fast enough.
 - Functionality to copy to clipboard the senders/recipient's ID on the transaction screen.
 - Optional icons for inputs.
+- Account signatures.
 
 ### Changed